fixing migrations

+++ packages/Utd/Room/Database/migrations/2023_11_04_102838_add_top_user_id_to_room.php
@@ -23,8 +23,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->dropIndex('rooms_top_user_id_index');
-            $table->dropColumn('top_user_id');
+            if (Schema::hasColumn('rooms', 'top_user_id')) {
+                $table->dropForeign(['top_user_id']);
+                $table->dropColumn('top_user_id');
+            }
         });
     }
 };
