+++ packages/Utd/Room/Database/migrations/2025_10_28_064622_add_additional_admin_to_rooms_table.php
@@ -20,7 +20,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->dropColumn('additional_admin');
+            if (Schema::hasColumn('rooms', 'additional_admin')) {
+                $table->dropColumn('additional_admin');
+            }
         });
     }
 };
